restructure and new tests
