--- tests/functional/adapter/test_ephemeral.py
@@ -1,11 +1,0 @@
-import pytest
-from dbt.tests.adapter.ephemeral.test_ephemeral import BaseEphemeralMulti
-from dbt.tests.util import run_dbt, check_relations_equal
-
-class TestEphemeralMultiExasol(BaseEphemeralMulti):
-
-    def test_ephemeral_multi_exasol(self, project):
-        run_dbt(["seed"])
-        results = run_dbt(["run"])
-        assert len(results) == 3
-        check_relations_equal(project.adapter, ["SEED", "DEPENDENT", "DOUBLE_DEPENDENT", "SUPER_DEPENDENT"])
